end of Section 6

+++ src/Controller/MicroPostController.php
@@ -12,13 +12,18 @@ use App\Entity\MicroPost;
 use App\Form\MicroPostType;
 use App\Repository\MicroPostRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Psr\Log\LoggerInterface;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Session\Flash\FlashBagInterface;
+use Symfony\Component\HttpKernel\Exception\UnauthorizedHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Routing\RouterInterface;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 
 /**
  * @Route("/micro-post")
@@ -50,6 +55,14 @@ class MicroPostController
      * @var FlashBagInterface
      */
     private $flashBag;
+    /**
+     * @var LoggerInterface
+     */
+    private $logger;
+    /**
+     * @var AuthorizationCheckerInterface
+     */
+    private $authorizationChecker;
 
     public function __construct(
         \Twig_Environment $twig,
@@ -57,7 +70,9 @@ class MicroPostController
         FormFactoryInterface $formFactory,
         EntityManagerInterface $entityManager,
         RouterInterface $router,
-        FlashBagInterface $flashBag)
+        FlashBagInterface $flashBag,
+        LoggerInterface $logger,
+        AuthorizationCheckerInterface $authorizationChecker)
     {
         $this->twig = $twig;
         $this->microPostRepository = $microPostRepository;
@@ -65,6 +80,8 @@ class MicroPostController
         $this->entityManager = $entityManager;
         $this->router = $router;
         $this->flashBag = $flashBag;
+        $this->logger = $logger;
+        $this->authorizationChecker = $authorizationChecker;
     }
 
     /**
@@ -72,6 +89,7 @@ class MicroPostController
      */
     public function index()
     {
+        /** @noinspection PhpUnhandledExceptionInspection */
         $html = $this->twig->render('micro-post/index.html.twig', [
             'posts' => $this->microPostRepository->findBy([], ['time' => 'DESC']),
         ]);
@@ -81,9 +99,14 @@ class MicroPostController
 
     /**
      * @Route("/edit/{id}", name="micro_post_edit")
+     * @Security("is_granted('edit',micropost)",message="access denied")
      */
     public function edit(MicroPost $micropost, Request $request)
     {
+        //$this->denyUnlessGranted('edit', $micropost);
+        if (!$this->authorizationChecker->isGranted('edit',$micropost)){
+            throw new UnauthorizedHttpException();
+        }
         $form = $this->formFactory->create(MicroPostType::class, $micropost);
         $form->handleRequest($request);
 
@@ -92,7 +115,7 @@ class MicroPostController
 
             return new RedirectResponse($this->router->generate('micro_post_index'));
         }
-
+        /** @noinspection PhpUnhandledExceptionInspection */
         return new Response(
             $this->twig->render('micro-post/add.html.twig', [
                 'form' => $form->createView()
@@ -103,9 +126,13 @@ class MicroPostController
 
     /**
      * @Route("/delete/{id}", name="micro_post_delete")
+     * @Security("is_granted('delete',micropost)",message="access denied")
      */
     public function delete(MicroPost $micropost)
     {
+        if (!$this->authorizationChecker->isGranted('delete',$micropost)){
+            throw new UnauthorizedHttpException();
+        }
         $this->entityManager->remove($micropost);
         $this->entityManager->flush();
 
@@ -116,12 +143,15 @@ class MicroPostController
 
     /**
      * @Route("/add", name="micro_post_add")
+     * @Security("is_granted('ROLE_USER')")
      */
-    public function add(Request $request)
+    public function add(Request $request,TokenStorageInterface $tokenStorage)
     {
+        $user = $tokenStorage->getToken()->getUser();
+
         $micropost = new MicroPost();
         $micropost->setTime(new \DateTime());
-
+        $micropost->setUser($user);
         $form = $this->formFactory->create(MicroPostType::class, $micropost);
         $form->handleRequest($request);
 
@@ -132,6 +162,7 @@ class MicroPostController
             return new RedirectResponse($this->router->generate('micro_post_index'));
         }
 
+        /** @noinspection PhpUnhandledExceptionInspection */
         return new Response(
             $this->twig->render('micro-post/add.html.twig', [
                 'form' => $form->createView()
@@ -144,6 +175,7 @@ class MicroPostController
      */
     public function post(MicroPost $post)
     {
+        /** @noinspection PhpUnhandledExceptionInspection */
         return new Response(
             $this->twig->render('micro-post/post.html.twig', [
                 'post' => $post,
